Fix compiler warnings.

+++ src/humlib.cpp
@@ -1,7 +1,7 @@
 //
 // Programmer:    Craig Stuart Sapp <craig@ccrma.stanford.edu>
 // Creation Date: Sat Aug  8 12:24:49 PDT 2015
-// Last Modified: Fri Sep  6 14:47:28 PDT 2019
+// Last Modified: Fri Sep  6 22:56:51 PDT 2019
 // Filename:      /include/humlib.cpp
 // URL:           https://github.com/craigsapp/humlib/blob/master/src/humlib.cpp
 // Syntax:        C++11
@@ -46096,11 +46096,11 @@ void Tool_imitation::analyzeImitation(vector<vector<string>>& results,
 
 			if (m_mark) {
 				for (int z=0; z<count; z++) {
-					if (i+z >= attacks.at(v1).size()) {
+					if (i+z >= (int)attacks.at(v1).size()) {
 						break;
 					}
 					token1 = attacks.at(v1).at(i+z)->getToken();
-					if (j+z >= attacks.at(v2).size()) {
+					if (j+z >= (int)attacks.at(v2).size()) {
 						break;
 					}
 					token2 = attacks.at(v2).at(j+z)->getToken();
